requirements.txt
qrcode 
colorama
pyinstaller
pillow
customtkinter
packaging